+++ CHANGELOG.md
@@ -5,6 +5,7 @@ _See also: [the official libGDX changelog](https://github.com/libgdx/libgdx/blob
 - **[UPDATE]** Updated to libGDX 1.12.1.
 - **[UPDATE]** Updated to Kotlin 1.9.20.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.7.3.
+- **[UPDATE]** Updated to VisUI 1.5.3.
 - **[UPDATE]** Updated to Dokka 1.9.10.
 
 #### 1.12.0-rc1