+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 _See also: [the official libGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
-#### 1.12.1-SNAPSHOT
+#### 1.12.1-rc1
 
 - **[UPDATE]** Updated to libGDX 1.12.1.
 - **[UPDATE]** Updated to Kotlin 1.9.20.